==> ClassificationExamples/iris/notebooks/4_Train_Models.ipynb <==
# Iris Classification

## Model Training

Loads data, instantiates model, and runs fit method. Saves model using joblib library. 

**NOTE:** Many sklearn models adhere to the same method: instantiate, fit, predict.

import os
import pandas as pd
from joblib import dump
import matplotlib.pyplot as plt

os.getcwd()

data_path = os.path.join('..', 'data')

X_train_data_path = os.path.join(data_path, 'X_train.csv')
y_train_data_path = os.path.join(data_path, 'y_train.csv')

X_train = pd.read_csv(X_train_data_path)
y_train = pd.read_csv(y_train_data_path)

from sklearn.ensemble import RandomForestClassifier

rf = RandomForestClassifier()

rf.fit(X_train, y_train.target)

models_path = os.path.join('..', 'models')
os.makedirs(models_path, exist_ok=True)
rf_model_path = os.path.join(models_path, 'rf.pkl')

dump(rf, rf_model_path)

from sklearn.linear_model import LogisticRegression

lr = LogisticRegression(max_iter=1000)

lr.fit(X_train, y_train.target)

lr_model_path = os.path.join(models_path, 'lr.pkl')
dump(lr, lr_model_path)
