==> ClassificationExamples/iris/notebooks/5_Predict_and_Evaluate.ipynb <==
# Iris Classification

## Model evaluation

Many metrics can be used to determine the quality of a classification model.

Evaluation should be decided based on test data performance. However, stats on the training data may also be helpful for determining under or over fitting.

This notebook will run metrics on both datasets, but **test metrics** are the primary focus to evaluate model generalizability.

import os
import pandas as pd
import seaborn as sns
from joblib import load

data_path = os.path.join('..', 'data')
# training data
X_train_path = os.path.join(data_path, 'X_train.csv')
y_train_path = os.path.join(data_path, 'y_train.csv')
# test data
X_test_path = os.path.join(data_path, 'X_test.csv')
y_test_path = os.path.join(data_path, 'y_test.csv')

X_train = pd.read_csv(X_train_path)
y_train = pd.read_csv(y_train_path)

X_test = pd.read_csv(X_test_path)
y_test = pd.read_csv(y_test_path)

models_path = os.path.join('..', 'models')

lr_model_path = os.path.join(models_path, 'lr.pkl')
lr = load(lr_model_path)
lr_train_preds = lr.predict(X_train)
lr_test_preds = lr.predict(X_test)

lr_test_preds

rf_model_path = os.path.join(models_path, 'rf.pkl')
rf = load(rf_model_path)
rf_train_preds = rf.predict(X_train)
rf_test_preds = rf.predict(X_test)

rf_test_preds

images_path = os.path.join('..', 'imgs')
os.makedirs(images_path, exist_ok=True)

lr_train_cm_path = os.path.join(images_path, 'lr_train_cm.png')
rf_train_cm_path = os.path.join(images_path, 'rf_train_cm.png')

lr_test_cm_path = os.path.join(images_path, 'lr_test_cm.png')
rf_test_cm_path = os.path.join(images_path, 'rf_test_cm.png')

from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.metrics import ConfusionMatrixDisplay
import matplotlib.pyplot as plt

# create list of preds and paths
true_list = [y_train, y_test, y_train, y_test]
pred_list = [lr_train_preds, lr_test_preds, rf_train_preds, rf_test_preds]
path_list = [lr_train_cm_path, lr_test_cm_path, rf_train_cm_path, rf_test_cm_path]

for trues, preds, path in zip(true_list, pred_list, path_list):
    cm_disp = ConfusionMatrixDisplay.from_predictions(y_true=trues, y_pred=preds, cmap='Blues')
    print(f'Saving image to path: {path}.')
    cm_disp.figure_.savefig(path)

print(f'Logistic Regression Accuracy Score: {accuracy_score(y_test.target, lr_test_preds): .2f} ')
print(f'Random Forest Accuracy Score: {accuracy_score(y_test.target, rf_test_preds): .2f} ')

print(classification_report(y_test.target, lr_test_preds))

print(classification_report(y_test.target, rf_test_preds))
